Drop Data::Visitor in favor of my own visitor code that does postorder. Trying to think if there's ever a time when it would be really important to have preorder instead; I could add it but I don't want to clutter up the interface unnecessarily. Left out for now.
Using the Visitor plugin (and depending on postorder traversal) is the
ActivePath plugin, which determines a path of "active" menu entries down to
the one that most accurately matches the user's current location.

Altered test.pl (which still isn't really a test) to use the subclassing
pattern that I intend to use for the real deal; the subclass provides the
interface to the outside app; in this case we mock the user's path for testing.



git-svn-id: svn+ssh://cleverdomain.org/publicsvn/WWW-MenuGrinder@3 08b0a27e-8b53-4eba-951b-c6525441d3d8