"Multi-pass" visitor support fixes ActivePath behavior.
ActivePath was potentially doing the wrong thing -- because it was
making its decision on "longest thing I've seen" before it processed
everything it could mark a node as active="yes" when it really wasn't
the best match, and also leave children with active, while their parents
didn't have active="child".

Now we do one pass to find the longest, and one pass to mark it and its
parents. It's actually simpler than it was before.

This is made possible by a new "multi pass" interface in
Plugin::Visitor; the new cleanup interface in MenuGrinder, and the new
plugin_verify interface which lets us make sure ahead of time that a
plugin hasn't declared some interface that it doesn't have.